imported and used yolo, also created a tracker with the metric of iou to decide if its a new id or not, its working but it could be changed depending on the new requirements
